made ids integer instead of long + added initial support for QA Step for Dataset

+++ src/client/gov/epa/emissions/framework/client/data/editor/EditorPanel.java
@@ -88,7 +88,7 @@ public class EditorPanel extends JPanel implements EditorPanelView {
     }
 
     private EditablePagePanel createEditablePage(Page page) {
-        editablePage = new EditablePage((int) dataset.getId(), version, page, tableMetadata);
+        editablePage = new EditablePage(dataset.getId(), version, page, tableMetadata);
         editablePagePanel = new EditablePagePanel(editablePage, messagePanel, changeablesList);
 
         return editablePagePanel;
