+++ src/contract/gov/epa/emissions/framework/services/data/DataCommonsService.java
@@ -59,11 +59,11 @@ public interface DataCommonsService {
     Status[] getStatuses(String username) throws EmfException;
 
     // Revisions
-    Revision[] getRevisions(long datasetId) throws EmfException;
+    Revision[] getRevisions(int datasetId) throws EmfException;
     void addRevision(Revision revision) throws EmfException;
 
     // Notes
-    Note[] getNotes(long datasetId) throws EmfException;
+    Note[] getNotes(int datasetId) throws EmfException;
     void addNote(Note note) throws EmfException;
     void addNotes(Note[] notes) throws EmfException;
